+++ src/components/NewProject/editapi/index.js
@@ -1,9 +1,7 @@
 import React from "react";
 import { CssBaseline, Container } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
-import TextInput from "./TextInput";
-import SimpleSelect from "./SimpleSelect";
-import Version from "./Version";
+import TextInput from "../TextInput";
 
 const useStyles = makeStyles((theme) => ({
   paper: {
@@ -13,10 +11,11 @@ const useStyles = makeStyles((theme) => ({
   },
   form: {
     width: "100%", // Fix IE 11 issue.
+    marginTop: theme.spacing(1),
   },
 }));
 
-export default function CreateApi() {
+export default function EditApi() {
   const classes = useStyles();
 
   return (
@@ -26,12 +25,7 @@ export default function CreateApi() {
         <form className={classes.form} noValidate>
           <TextInput name="apiName" label="API Name" />
           <TextInput name="endpoint" label="Endpoint (URL)" />
-          <SimpleSelect
-            title="API Type"
-            options={["GET", "POST", "PUT", "DELETE"]}
-          />
-          {/* VERSIONS */}
-          <Version />
+          <TextInput name="apiType" label="API Type" />
         </form>
       </div>
     </Container>
