+++ src/pages/content/home.js
@@ -1,0 +1,30 @@
+export const projects = [
+  {
+    title: "Project 1",
+    subtitle: "adjective",
+    description: "lorem ipsum",
+    link: "htts://google.com",
+    linkText: "Go to link",
+  },
+  {
+    title: "Project 2",
+    subtitle: "adjective",
+    description: "lorem ipsum",
+    link: "htts://google.com",
+    linkText: "Go to link",
+  },
+  {
+    title: "Project 3",
+    subtitle: "adjective",
+    description: "lorem ipsum",
+    link: "htts://google.com",
+    linkText: "Go to link",
+  },
+  {
+    title: "Project 4",
+    subtitle: "adjective",
+    description: "lorem ipsum",
+    link: "htts://google.com",
+    linkText: "Go to link",
+  },
+];
