+++ src/pages/home.js
@@ -1,34 +1,29 @@
 import React from "react";
-import { createStyles, makeStyles } from "@material-ui/core/styles";
-// import { Fab } from "@material-ui/core";
-// import AddIcon from "@material-ui/icons/Add";
-// import { useHistory } from "react-router-dom";
-import ResponsiveDrawer from "../components/ResponsiveDrawer";
-import Container from "../components/Home/Container";
+import { makeStyles } from "@material-ui/core/styles";
+import { Grid } from "@material-ui/core";
+import Card from "../components/Home/Card";
+import CardAdd from "../components/Home/CardAdd";
 
-const useStyles = makeStyles((theme) =>
-  createStyles({
-    root: {
-      "& > *": {
-        margin: theme.spacing(1),
-      },
-    },
-  })
-);
+import homeStyles from "./styles/home";
+import { projects } from "./content/home";
+
+const useStyles = makeStyles(homeStyles);
 
 export default function Home() {
   const classes = useStyles();
-  // let history = useHistory();
-
-  // const handleNext = () => {
-  //   history.push("/signin");
-  // };
 
   return (
     <div className={classes.root}>
-      <ResponsiveDrawer title="Amiba Dashboard">
-        <Container />
-      </ResponsiveDrawer>
+      <Grid container spacing={3}>
+        {projects.map((item, index) => (
+          <Grid item xs={12} sm={6} md={3} lg={4} key={item.title}>
+            <Card {...item} />
+          </Grid>
+        ))}
+        <Grid item xs={12} sm={6} md={3} lg={4}>
+          <CardAdd />
+        </Grid>
+      </Grid>
     </div>
   );
 }
