+++ src/pages/styles/home.js
@@ -1,0 +1,12 @@
+export default function homeStyles(theme) {
+  return {
+    root: {
+      flexGrow: 1,
+    },
+    paper: {
+      padding: theme.spacing(2),
+      textAlign: "center",
+      color: theme.palette.text.secondary,
+    },
+  };
+}
