+++ src/routes.js
@@ -1,9 +1,11 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+
 import SignUp from "./pages/signup";
 import SignIn from "./pages/signin";
 import Home from "./pages/home";
 import NewProject from "./pages/newproject";
+import ResponsiveDrawer from "./components/ResponsiveDrawer";
 
 function Routes() {
   return (
@@ -16,12 +18,14 @@ function Routes() {
           <Route exact path="/signup">
             <SignUp />
           </Route>
-          <Route exact path="/newproject">
-            <NewProject />
-          </Route>
-          <Route exact path="/">
-            <Home />
-          </Route>
+          <ResponsiveDrawer>
+            <Route exact path="/newproject">
+              <NewProject />
+            </Route>
+            <Route exact path="/">
+              <Home />
+            </Route>
+          </ResponsiveDrawer>
         </Switch>
       </div>
     </Router>
